+++ app/assets/config/manifest.js
@@ -2,3 +2,5 @@
 //= link application.js
 //= link_tree ../../../vendor/assets/images
 //= link_tree ../builds
+//= link_tree ../../javascript .js
+//= link_tree ../../../vendor/javascript .js
