+++ app/javascript/application.js
@@ -1,0 +1,3 @@
+// Configure your import map in config/importmap.rb. Read more: https://github.com/rails/importmap-rails
+import * as $ from 'jquery';
+import "controllers"
