+++ app/javascript/controllers/popup_nav_controller.js
@@ -1,0 +1,23 @@
+import { Controller } from "@hotwired/stimulus"
+
+export default class extends Controller {
+  connect() {
+    var arrowIcon        = $('.header__popup-link');
+    var popupNav         = $('.header__popup__nav-links');
+
+    var navExpandedClass = 'is-expanded';
+
+    function removeNavExpandedClass() {
+      popupNav.removeClass(navExpandedClass);
+    }
+
+    function addNavExpandedClass() {
+      popupNav.addClass(navExpandedClass);
+    }
+
+    arrowIcon.click(function(e){
+      var nav = {expandedClass: navExpandedClass, popUp: popupNav}
+      handleClick(e, nav, removeNavExpandedClass, addNavExpandedClass);
+    });
+  }
+}
