+++ app/javascript/handle_click.js
@@ -4,7 +4,6 @@
 // It's not advisable to add code directly here, but if you do, it'll appear at the bottom of the
 // the compiled file.
 //
-//= require jquery3
 //= require jquery_ujs
 //= require clipboard
 //= require github_buttons
