+++ app/javascript/jquery_tmp.js
@@ -1,0 +1,4 @@
+// This is needed to add jquery to the global scope
+import jquery from "jquery"
+window.jQuery = jquery
+window.$ = jquery
